refactor: create validators
