+++ src/main/kotlin/lotto/validators/ValidationErrors.kt
@@ -1,0 +1,8 @@
+package lotto.validators
+
+enum class ValidationErrors(val message: String) {
+    MUST_DIVIDE_BY_TICKET_PRICE("The amount must be divisible by 1,000"),
+    MUST_BE_IN_RANGE("Bonus number must be between 1 and 45."),
+    MUST_ENTER_SIX_NUMBERS("You need to write 6 comma-separated numbers"),
+    MUST_ENTER_UNIQUE_NUMBERS("You need to write 6 unique numbers"),
+}
